Always throw exceptions on SQL failures
We definitely want the default to be that an exception gets thrown if there
are any problems executing an SQL statement. If we need to ignore an
exception we can just catch it. Otherwise problems that happen because a
query fails don't get caught until way after the problem happens.