Now create automatically runs config if amp not configured.
When runing 'amp create' from inside another script it was erroring out
because I hadn't already run 'amp config'. This checks to see if the
mysql_dsn variable has been configured and if not, then it runs the 'amp
config' command which prompts for those vars.

I had to add an option to the create and export commands to let you output
the amp environment vars to a separate file to separate the prompts to
configure amp from the environment variable output.