InstanceRepository - Provide better hints for DB names
The naming convention for databases is based on the folder-naming -- i.e.
the folder-name is a "hint" that becomes the start of the DB name.

However, in a file-naming structure like the following, all the folders produce the same "hint":

* `/var/www/foo/web`
* `/var/www/bar/web`
* `/var/www/whiz/web`
* `/var/www/bang/web`

This patch changes the hinting formula to disregard common endings (`/web`,
`/htdocs`, `/www`).